Chore: start refactoring variants logic into its own package
